Update scrapeScenes.py
Fixed an error in the code found by me (and fixed by bnkai) where the script was getting a "Scrape succeeded, but update failed" error when trying to handle a site performer which differs from the parent performer (ie. Lea Lexis site performer and Lea Lexus parent performer). This code change has been tested and now scraping as it's supposed to.