Switch WaitForIdle to Queue system
In line with latest changes to Stash dev, JobStatus is no longer a GraphQL query. Instead, we need to check if a queue exists for jobs when checking if Stash is idle. Also removed WaitForIdle on individual tasks as those can be queued properly at Stash now, but kept the wait flag, so users can choose to wait for the queue to be completed in their workflows