Putting in a -p/--path_include option for path masking
On a system with ~45k scenes and 1.5k tags, the regular run was taking almost 5 minutes per page of 100 results.  This was due to searching for scenes without the "Scraped from TPDB" tag.  With the path_include you can direct the scraper to only search for files with a given string in the filepath (path, filename, etc...  anywhere in the path) to limit that query run time.

By way of example on that same system with ~45k scenes, the import directory had ~1.2k scenes.  When pointed to the script would take only around 10 seconds per page of 100 results.

Plus a few formatting tweaks before I gave up on it.  There are a lot of Lint errors in this thing, but nothing to keep anything from running correctly